+++ features/vaadin-dashlets/dashlet-summary/src/main/java/org/opennms/features/vaadin/dashboard/dashlets/SummaryDashlet.java
@@ -28,15 +28,26 @@
 
 package org.opennms.features.vaadin.dashboard.dashlets;
 
-import com.vaadin.server.Page;
-import com.vaadin.server.Sizeable;
-import com.vaadin.server.ThemeResource;
-import com.vaadin.ui.*;
 import org.opennms.core.criteria.CriteriaBuilder;
-import org.opennms.features.vaadin.dashboard.model.*;
+import org.opennms.features.vaadin.dashboard.model.AbstractDashlet;
+import org.opennms.features.vaadin.dashboard.model.AbstractDashletComponent;
+import org.opennms.features.vaadin.dashboard.model.Dashlet;
+import org.opennms.features.vaadin.dashboard.model.DashletComponent;
+import org.opennms.features.vaadin.dashboard.model.DashletSpec;
 import org.opennms.netmgt.dao.api.AlarmDao;
 import org.opennms.netmgt.model.OnmsAlarm;
 import org.opennms.netmgt.model.OnmsSeverity;
+
+import com.vaadin.server.Page;
+import com.vaadin.server.Sizeable;
+import com.vaadin.server.ThemeResource;
+import com.vaadin.ui.Accordion;
+import com.vaadin.ui.Alignment;
+import com.vaadin.ui.Component;
+import com.vaadin.ui.HorizontalLayout;
+import com.vaadin.ui.Image;
+import com.vaadin.ui.Label;
+import com.vaadin.ui.VerticalLayout;
 
 /**
  * This class implements a {@link Dashlet} for testing purposes.
@@ -137,7 +148,7 @@ public class SummaryDashlet extends AbstractDashlet {
      * @param notAck number of unacknowledged alarms
      * @return the trend value
      */
-    private int computeTrend(int ack, int notAck) {
+    private int computeTrend(long ack, long notAck) {
         if (ack == notAck) {
             return TREND_EAST;
         } else {
@@ -223,8 +234,8 @@ public class SummaryDashlet extends AbstractDashlet {
             horizontalLayout.addStyleName("summary");
             horizontalLayout.addStyleName(onmsSeverity.name().toLowerCase());
 
-            int acknowledged = countBySeverity(true, m_timeslot, onmsSeverity);
-            int notAcknowledged = countBySeverity(false, m_timeslot, onmsSeverity);
+            long acknowledged = countBySeverity(true, m_timeslot, onmsSeverity);
+            long notAcknowledged = countBySeverity(false, m_timeslot, onmsSeverity);
 
             Label labelSeverity = new Label(onmsSeverity.getLabel());
             labelSeverity.addStyleName("summary-font");
@@ -326,8 +337,8 @@ public class SummaryDashlet extends AbstractDashlet {
                 }
             }
 
-            int acknowledged = countByUei(true, m_timeslot, uei);
-            int notAcknowledged = countByUei(false, m_timeslot, uei);
+            long acknowledged = countByUei(true, m_timeslot, uei);
+            long notAcknowledged = countByUei(false, m_timeslot, uei);
 
             Label labelSeverity = new Label(uei.replace("uei.opennms.org/nodes/", ""));
             labelSeverity.addStyleName("summary-font");
@@ -406,7 +417,7 @@ public class SummaryDashlet extends AbstractDashlet {
      * @param onmsSeverity the {@link OnmsSeverity} to search for
      * @return number of alarms found
      */
-    public int countBySeverity(boolean acknowledged, long age, OnmsSeverity onmsSeverity) {
+    public long countBySeverity(boolean acknowledged, long age, OnmsSeverity onmsSeverity) {
         CriteriaBuilder criteriaBuilder = new CriteriaBuilder(OnmsAlarm.class);
 
         if (acknowledged) {
@@ -430,7 +441,7 @@ public class SummaryDashlet extends AbstractDashlet {
      * @param uei          search for alarms with the specified uei
      * @return number of alarms found
      */
-    public int countByUei(boolean acknowledged, long age, String uei) {
+    public long countByUei(boolean acknowledged, long age, String uei) {
         CriteriaBuilder criteriaBuilder = new CriteriaBuilder(OnmsAlarm.class);
 
         if (acknowledged) {
