+++ opennms-webapp/src/main/java/org/opennms/web/outage/DaoWebOutageRepository.java
@@ -224,7 +224,7 @@ public class DaoWebOutageRepository implements WebOutageRepository, Initializing
     /** {@inheritDoc} */
     @Transactional
     @Override
-    public int countMatchingOutages(OutageCriteria criteria) {
+    public long countMatchingOutages(OutageCriteria criteria) {
         return m_outageDao.countMatching(getOnmsCriteria(criteria));
     }
