+++ opennms-webapp/src/main/java/org/opennms/web/rest/NotificationRestService.java
@@ -108,7 +108,7 @@ public class NotificationRestService extends OnmsRestService {
     public String getCount() {
         readLock();
         try {
-            return Integer.toString(m_notifDao.countAll());
+            return Long.toString(m_notifDao.countAll());
         } finally {
             readUnlock();
         }
