+++ opennms-webapp/src/test/java/org/opennms/web/svclayer/DefaultOutageServiceIntegrationTest.java
@@ -123,7 +123,7 @@ public class DefaultOutageServiceIntegrationTest implements InitializingBean {
     @Transactional
     @Ignore
     public void testNoOfSuppressedOutages(){
-        Integer outages = m_outageService.getSuppressedOutageCount();
+        Long outages = m_outageService.getSuppressedOutageCount();
         assertTrue("We should find suppressed messages ", outages == 0);
     }
